+++ drivers/virtio/virtqueue.c
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include <zephyr/virtio/virtqueue.h>
+#include <zephyr/drivers/virtio/virtqueue.h>
 #include <zephyr/kernel.h>
 #include <zephyr/logging/log.h>
 #include <zephyr/sys/__assert.h>
