+++ include/zephyr/drivers/virtio/virtqueue.h
